Big changes (read description)
Change of styles, many went from normal to tailwind.

We created a single sidebar with the necessary styles, the previous one was deleted.

Loading in registration and Login.

We created the membership creation, using the fetch to the endpoint.

We made the list of users (no matter if admin or client).

Redistributed the admin panel with folders.